--- 03_Selected_CVs_File/selected_resume.csv
@@ -1,4 +1,0 @@
-,Matched_CV,Score
-1,Resume_1.pdf,92.0
-2,Resume_11.pdf,92.0
-3,Resume_7.pdf,67.0
