--- description.txt
@@ -1,1 +1,0 @@
-team member - ujjwal,Ranjit, Abhishek, Milon
